refactor: move value aliases into its own file
